Refactor: Remove outdated chip_info and config_downloader workflows
- Deleted `update_props_pr.py` and `flow.py` from `chip_info` workflow as they are no longer needed.
- Removed `__init__.py` from `config_downloader` workflow.
- Cleaned up `device_gateway_integration` workflow by deleting unnecessary files and consolidating functionality into a new `gateway_integration` flow.
- Added new `gateway_integration` flow and tasks to handle device topology requests and responses.
- Introduced a new scheduler flow to manage execution IDs and scheduling tasks.
- Updated `chip_report` flow to include Slack notifications for successful report generation.
