--- src/qdash/workflow/worker/device_gateway_integration/__init__.py
@@ -1,4 +1,0 @@
-from qdash.workflow.worker.chip_info.flow import update_props
-from qdash.workflow.worker.chip_info.update_props_pr import update_props_pr
-
-__all__ = ["update_props", "update_props_pr"]
